event_platform/requirements.txt
fastapi
uvicorn
motor
python-jose
python-multipart